renamed files in accordance with standards
