+++ public/src/components/script/script_list.jsx
@@ -1,22 +1,21 @@
 import React, { Component } from 'react';
 import { Table } from 'react-bootstrap';
 
-import QuestionTableRow from './questions_table_row';
+import ScriptsTableRow from './script_table_row';
 import AdminDashboardButtonGroup from '../common/admin_nav_btn_group';
 import AdminBanner from '../common/admin_welcome_banner';
 import CreateNewButton from '../common/admin_createNew_btn';
 import buttons_obj from '../common/admin_button_objs';
 
-export default class QuestionsList extends Component {
+export default class ScriptsList extends Component {
   componentDidMount() {
-    this.props.fetchAllQuestions();
+    this.props.fetchAllScripts();
   }
 
   render() {
-    const { all_questions, account_info: { last_name, first_name }, history } = this.props;
-    const thisPage = 'Question';
+    const { all_scripts, account_info: { last_name, first_name }, history } = this.props;
+    const thisPage = 'Script';
     const { view_edit } = buttons_obj;
-
     return (
       <div>
         <AdminBanner
@@ -26,15 +25,14 @@ export default class QuestionsList extends Component {
           page={thisPage}
         />
         <div>
-          <AdminDashboardButtonGroup history={history} />
+          <AdminDashboardButtonGroup history={this.props.history} />
         </div>
         <Table responsive>
           <thead>
             <tr>
-              <th>Title</th>
+              <th>Name</th>
+              <th>Body</th>
               <th>Description</th>
-              <th>Type</th>
-              <th>Response Options</th>
               <th>Created At</th>
               <th>Updated At</th>
               <th />
@@ -42,13 +40,13 @@ export default class QuestionsList extends Component {
           </thead>
 
           <tbody>
-            {all_questions && all_questions.length ?
-              all_questions.map(question =>
+            {all_scripts && all_scripts.length ?
+              all_scripts.map(script =>
                 (
-                  <QuestionTableRow
-                    key={question.id}
-                    question={question}
-                    handleEditClick={this.props.setCurrentQuestion}
+                  <ScriptsTableRow
+                    key={script.id}
+                    script={script}
+                    handleEditClick={this.props.setCurrentScript}
                     buttons={view_edit}
                     page={thisPage}
                   />
@@ -59,7 +57,7 @@ export default class QuestionsList extends Component {
           </tbody>
         </Table>
         <div>
-          <CreateNewButton {...this.props} page={thisPage} path={'/admin/questions/new'} />
+          <CreateNewButton {...this.props} page={thisPage} path={'/admin/scripts/new'} />
         </div>
       </div>
     );
