+++ public/src/components/script/script_table_row.jsx
