+++ public/src/containers/list_campaigns.jsx
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 import { withRouter } from 'react-router-dom';
 
-import CampaignList from '../components/campaign_list';
+import CampaignList from '../components/campaign/campaign_list';
 
 import { fetchAllCampaigns, setCurrentCampaign } from '../actions/campaign';
